fix: DHIS2-10054: Code review fixes.
